Not sure how I lost it, but I lost the original git repo. This is the new initial commit.
Getting ready to switch to Angular.js as the framework. Once I learn how to use it correctly,
it should make working with the multiple views thing a snap and keep the app a one-page web app.

I may even be able to get rid of jQuery, but one thing at a time.